move related tests to separate file. move mux to search struct
